docker-deploy/web-app/requirements.txt
Django>=2.0,<3.0
psycopg2
django-crispy-forms==1.8.1
Pillow
protobuf
